Split LoginRegister into separate Login, Register, and ForgotPassword pages
