+++ src/components/Router.jsx
@@ -6,7 +6,9 @@ import {
 } from "react-router-dom";
 
 import Landing from "../pages/Landing";
-import LoginAndRegister from "../pages/LoginRegister";
+import Login from "../pages/Login";
+import Register from "../pages/Register";
+import ForgotPassword from "../pages/ForgotPassword";
 import ResetPassword from "../pages/ResetPassword";
 import PostCreate from "../pages/PostCreate";
 import Post from "../pages/Post";
@@ -22,7 +24,9 @@ export default function AppRouter() {
     <Router>
             <Routes>
               <Route index element={<Landing />} />
-              <Route path="/login" element={<LoginAndRegister />} />
+              <Route path="/login" element={<Login />} />
+              <Route path="/register" element={<Register />} />
+              <Route path="/forgot-password" element={<ForgotPassword />} />
               <Route path="/reset-password" element={<ResetPassword />} />
 
               <Route
